refactor : rename Player to PlayerStreaming
